%ld and time_t don't match, so cast the argument to long to ease portability problems
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@47295 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_queue.c
@@ -2620,7 +2620,7 @@ static int try_calling(struct queue_ent *qe, const char *options, char *announce
 		/* use  pbx_builtin_setvar to set a load of variables with one call */
 		if (qe->parent->setinterfacevar) {
 			snprintf(interfacevar,sizeof(interfacevar), "MEMBERINTERFACE=%s|MEMBERNAME=%s|MEMBERCALLS=%d|MEMBERLASTCALL=%ld|MEMBERPENALTY=%d|MEMBERDYNAMIC=%d",
-				member->interface, member->membername, member->calls, member->lastcall, member->penalty, member->dynamic);
+				member->interface, member->membername, member->calls, (long)member->lastcall, member->penalty, member->dynamic);
 		 	pbx_builtin_setvar(qe->chan, interfacevar);
 		}
